requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
python-dotenv==1.0.1
pymysql==1.1.0
